break values up into smaller modules
