+++ mica-language/src/bytecode.rs
@@ -129,6 +129,9 @@ pub enum Opcode {
    CreateStruct,
    /// Creates a list from `operand` values that are at the top of the stack.
    CreateList,
+   /// Creates a dict from `operand * 2` values that are at the top of the stack. The values have
+   /// to be arranged in `key, value, key, value...` order, from bottom to top.
+   CreateDict,
 
    /// Assigns the value at the top of the stack to a global. The value stays on the stack.
    AssignGlobal,
