+++ mica-language/src/common.rs
@@ -83,6 +83,7 @@ pub enum ErrorKind {
    LeftParenExpected,
    UnexpectedEof,
    CommaExpected,
+   ColonExpectedAfterDictKey,
 
    // Code generator
    VariableDoesNotExist(Rc<str>),
@@ -110,6 +111,7 @@ pub enum ErrorKind {
    FieldOutsideOfImpl,
    MissingFields(Vec<Rc<str>>),
    ListIsTooLong,
+   DictIsTooLarge,
 
    // Runtime
    TypeError {
@@ -164,6 +166,7 @@ impl std::fmt::Display for ErrorKind {
          Self::LeftParenExpected => write!(f, "left parenthesis '(' expected"),
          Self::UnexpectedEof => write!(f, "unexpected end of file"),
          Self::CommaExpected => write!(f, "comma ',' expected"),
+         Self::ColonExpectedAfterDictKey => write!(f, "colon ':' expected after dict key"),
 
          Self::VariableDoesNotExist(name) => write!(f, "variable '{name}' does not exist"),
          Self::InvalidAssignment => write!(f, "invalid left hand side of assignment"),
@@ -204,6 +207,7 @@ impl std::fmt::Display for ErrorKind {
             )
          }
          Self::ListIsTooLong => write!(f, "list literal has too many elements"),
+         Self::DictIsTooLarge => write!(f, "dict literal has too many pairs"),
 
          Self::TypeError { expected, got } => {
             write!(f, "type mismatch, expected {expected} but got {got}")
